+++ src/util/buffer.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-static Buffer default_alloc(size_t size, const char *filename, int line) {
+static Buffer default_alloc(uint32_t size, const char *filename, int line) {
   return malloc(size);
 }
 
@@ -13,10 +13,11 @@ void Buffer_init(Buffer_Alloc alloc, Buffer_Free free) {
   g_alloc = alloc;
   g_free = free;
 }
-Buffer Buffer_alloc_(size_t size, const char *filename, int line) {
+Buffer Buffer_alloc_(uint32_t size, const char *filename, int line) {
   return g_alloc(size, filename, line);
 }
-Buffer Buffer_from_(size_t size, Buffer source, const char *filename, int line) {
+Buffer Buffer_from_(uint32_t size, Buffer source, const char *filename,
+                    int line) {
   void *buf = g_alloc(size, filename, line);
   memcpy(buf, source, size);
   return buf;
